[IMP] travis emulator enhancements

+++ clodoo/run_odoo_debug.sh
@@ -16,20 +16,20 @@ if [ -z "$Z0LIBDIR" ]; then
   exit 2
 fi
 
-__version__=0.1.7
+__version__=0.1.8
 
 OPTOPTS=(h        d        m           n            t         s        U         V           v           x)
 OPTDEST=(opt_help opt_db   opt_modules opt_dry_run  opt_touch opt_stop opt_user  opt_version opt_verbose opt_xport)
 OPTACTI=(1        "="      "="         "1"          1         1        "="       "*>"        1           "=")
-OPTDEFL=(1        ""       ""          0            0         0        "postgres" ""          0           "")
-OPTMETA=("help"   "dbname" "modules"   "do nothing" "touch"   ""       "user"     "version"   "verbose"   "port")
+OPTDEFL=(1        ""       ""          0            0         0        ""        ""          0           "")
+OPTMETA=("help"   "dbname" "modules"   "do nothing" "touch"   ""       "user"    "version"   "verbose"   "port")
 OPTHELP=("this help"\
  "db name to test"\
  "modules to test"\
  "do nothing (dry-run)"\
  "touch config file, do not run odoo"\
  "stop after init"\
- "db username (def postgres)"\
+ "db username"\
  "show version"\
  "verbose mode"\
  "set odoo xmlrpc port")
@@ -71,22 +71,39 @@ fi
 create_db=0
 drop_db=0
 if [ -n "$opt_modules" ]; then
-   opts="-i $opt_modules --test-enable"
-   create_db=1
-   if [ $opt_stop -gt 0 ]; then
-    opts="$opts --stop-after-init"
-   fi
-   if [ -z "$opt_db" ]; then
-     opt_db="test_openerp"
-     if [ $opt_stop -gt 0 ]; then
-       drop_db=1
-       if [ -z "$opt_xport" ]; then
-         opt_xport=807$sfxver
-       fi
-     fi
-   fi
+  mods=${opt_modules//,/ }
+  for m in $mods; do
+    p=$(find /opt/odoo/$odoo_ver$sfx -type d -name $m|head -n1)
+    if [ -f $p/__openerp__.py ]; then
+      f=$p/__openerp__.py
+      x=$(cat $f|grep -A10 depends|tr -d '\n'|awk -F"[" '{print $2}'|awk -F"]" '{print $1}'|tr -d '" '|tr -d "'")
+      if [ -n "$x" ]; then
+        opt_modules="$opt_modules,$x"
+      fi
+    fi
+  done
+  opts="-i $opt_modules --test-enable"
+  create_db=1
+  if [ $opt_stop -gt 0 ]; then
+    opts="$opts --stop-after-init --test-commit"
+  fi
+  if [ -z "$opt_db" ]; then
+    opt_db="test_openerp"
+    if [ $opt_stop -gt 0 ]; then
+      drop_db=1
+    fi
+  fi
+  if [ -z "$opt_xport" ]; then
+    let opt_xport="8070+$sfxver"
+  fi
+  if [ -z "$opt_user" ]; then
+    opt_user=odoo$sfxver
+  fi
 else
-   opts=""
+  opts=""
+  if [ -z "$opt_user" ]; then
+    opt_user=postgres
+  fi
 fi
 if [ -n "$opt_db" ]; then
    opts="$opts -d $opt_db"
@@ -101,6 +118,9 @@ if [ $opt_dry_run -eq 0 ]; then
   if [ -n "$opt_xport" ]; then
     sed -ie "s:^xmlrpc_port *=.*:xmlrpc_port = $opt_xport:" ~/.openerp_serverrc
   fi
+  if [ -n "$opt_user" ]; then
+    sed -ie "s:^db_user *=.*:db_user = $opt_user:" ~/.openerp_serverrc
+  fi
   if [ $opt_verbose -gt 0 ]; then
     vim ~/.openerp_serverrc
   fi
@@ -108,9 +128,9 @@ fi
 if [ $opt_touch -eq 0 ]; then
   if [ $drop_db -gt 0 ]; then
     if [ $opt_verbose -gt 0 ]; then
-      echo "dropdb -U$opt_user --if-exists $opt_db"
+      echo "pg_db_active -a $opt_db; dropdb -U$opt_user --if-exists $opt_db"
     fi
-    dropdb -U$opt_user --if-exists $opt_db
+    pg_db_active -a $opt_db; dropdb -U$opt_user --if-exists $opt_db
   fi
   if [ $create_db -gt 0 ]; then
     if [ $opt_verbose -gt 0 ]; then
@@ -130,9 +150,11 @@ if [ $opt_touch -eq 0 ]; then
     eval $script $opts
   fi
   if [ $drop_db -gt 0 ]; then
-    if [ $opt_verbose -gt 0 ]; then
-      echo "dropdb -U$opt_user --if-exists $opt_db"
+    if [ -z "$opt_modules" -o $opt_stop -eq 0 ]; then
+      if [ $opt_verbose -gt 0 ]; then
+        echo "dropdb -U$opt_user --if-exists $opt_db"
+      fi
+      dropdb -U$opt_user --if-exists $opt_db
     fi
-    dropdb -U$opt_user --if-exists $opt_db
   fi
 fi
